+++ filters/kword/ascii/asciiexport.cc
@@ -376,12 +376,12 @@ void ProcessTextTag ( QDomNode    myNode,
 static void CreateMissingFormatData(QString &paraText, QValueList<FormatData> &paraFormatDataList)
 {
     QValueList<FormatData>::Iterator  paraFormatDataIt;
-    int lastPos=0; // last position
+    unsigned int lastPos=0; // last position
 
     paraFormatDataIt = paraFormatDataList.begin ();
     while (paraFormatDataIt != paraFormatDataList.end ())
     {
-        if ((*paraFormatDataIt).pos>lastPos)
+        if (static_cast<unsigned int>((*paraFormatDataIt).pos)>lastPos)
         {
             //We must add a FormatData
             FormatData formatData(lastPos,(*paraFormatDataIt).pos-lastPos);
